Add cbc cipher mode operation support
The encrypted message now can be any length. CBC cipher mode support
added. When sending a message, an IV is initialized to encrypt it using
cbc and aes-128. The IV is sent with the message.
To decrypt the message the node retireve the IV and decrypt.
Maximum length of msg = SERVER_BUFFER_SIZE - MSG_TYPE_SIZE - IV_SIZE

Signed-Off-By: Stefano Milani <stefano.milani96@gmail.com>